refactor book data into its own folder, add books prop for the Books component
