+++ src/pages/data/BookItems.js
@@ -1,0 +1,54 @@
+const BookItems = [
+    {
+      Title: "Lord of the Flies",
+      Author: "William Golding",
+      Description:
+        "Tells a powerful story about the tension between our impulses as humans to follow the rules and act in the best interest of others or to satisfy our own immediate desires.",
+      Src:
+        "https://s3.amazonaws.com/www.dan-murphy.com/public/lord-of-the-flies.jpg",
+      Href:
+        "https://www.amazon.com/Lord-Flies-William-Golding/dp/0399501487/ref=sr_1_1?crid=2H4S4GZM37AF7&dchild=1&keywords=lord+of+the+flies&qid=1609023157&sprefix=lord+of+the+flies%2Caps%2C175&sr=8-1",
+    },
+    {
+      Title: "Acting with Power",
+      Author: "Deborah Gruenfeld",
+      Description:
+        "We all have more power than we are aware of, but it is often misused and misunderstood. This book teaches you when to use power, how to leverage it, and why it is so important for personal and professional growth.",
+      Src:
+        "https://s3.amazonaws.com/www.dan-murphy.com/public/acting-with-power.jpg",
+      Href:
+        "https://www.amazon.com/Acting-Power-More-Powerful-Believe/dp/1101903953/ref=sr_1_1?crid=3UUIHIGQ7VCGV&dchild=1&keywords=acting+with+power&qid=1609023215&sprefix=acting+with%2Caps%2C167&sr=8-1",
+    },
+    {
+      Title: "The New Silk Roads: A New History of the World",
+      Author: "Peter Frankopan",
+      Description:
+        "Provides a thorough analysis of the forces that have caused empires to succeed and fail and challenges many of the assumptions we hold about history.",
+      Src:
+        "https://s3.amazonaws.com/www.dan-murphy.com/public/the-silk-roads.jpg",
+      Href:
+        "https://www.amazon.com/Silk-Roads-New-History-World/dp/1101912375/ref=sr_1_2?crid=1FJVPNE6EQRGK&dchild=1&keywords=the+new+silk+roads+by+peter+frankopan&qid=1609023328&sprefix=the+new+silk+roads%3A+a+new+%2Caps%2C162&sr=8-2",
+    },
+    {
+      Title: "The Everything Store",
+      Author: "Brad Stone",
+      Description:
+        "The incredible story behind the rise of Amazon, it's many acquisitions, AWS, and Jeff Bezos.",
+      Src:
+        "https://s3.amazonaws.com/www.dan-murphy.com/public/the-everything-store.jpg",
+      Href:
+        "https://www.amazon.com/Everything-Store-Jeff-Bezos-Amazon/dp/0316219282/ref=sr_1_1?crid=3NWX9UK39ILLF&dchild=1&keywords=the+everything+store&qid=1609023352&sprefix=the+everything+sto%2Caps%2C151&sr=8-1",
+    },
+    {
+      Title: "Animal Farm",
+      Author: "George Orwell",
+      Description:
+        "A novel about farm animals who rebel against their human farmer, hoping to create a society where the animals can be equal and free. This represents Russia and the Soviet Union under Communist Party rule.",
+      Src:
+        "https://s3.amazonaws.com/www.dan-murphy.com/public/animal-farm.jpg",
+      Href:
+        "https://www.amazon.com/Animal-Farm-George-Orwell/dp/0451526341/ref=sr_1_1?crid=DUGG67CDB8IZ&dchild=1&keywords=animal+farm+by+george+orwell&qid=1609020557&sprefix=animal+farm%2Caps%2C176&sr=8-1",
+    },
+  ];
+
+  export default BookItems;
